pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.1",
    "nest-asyncio",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
